first commit: add simple file storage commit
for this first commit, add a simple file storage backend for the coming
blobserver and define the blob.

A blob is a binary data with an a content id formatted as
<hashtype>-<contenthash>.

This storage store filesi on the local file system  in a directory given
as an option.

A storage backend is a module that provides the following functions:

- start_link/1 : start the backend and return its pid
- store/3 : store a blob with its id
- fetch/2 : fetch a blob
- delete/2 : delete a blob
- enumerate/3 : enumerate blobs over a function
- all/1 : return all blobs
- stat/2 : check if the given blobs are present or missing
- stat_fold/4: like state but pass blobs over a function.

Example of usage with a fake blob reference:

    1> {ok, Pid} = coffer_storage_file:start_link([{path,
<<"/tmp/coffer">>}]).
    {ok,<0.36.0>}
    2> coffer_storage_file:store(Pid, <<"test-hello">>, <<"hello">>).
    ok
    3> coffer_storage_file:fetch(Pid, <<"test-hello">>).
    {ok,<<"hello">>}
    4> coffer_storage_file:all(Pid).
    [{<<"test-hello">>,5}]